MINIFICPP-251 Move Civet implementations to an extension.
