+++ extensions/civet_curl_tests/unit/InvokeHTTPTests.cpp
@@ -25,13 +25,14 @@
 #include <set>
 #include "FlowController.h"
 #include "io/BaseStream.h"
-#include "../../TestBase.h"
+#include "TestBase.h"
 #include "processors/GetFile.h"
 #include "core/Core.h"
 #include "HTTPClient.h"
 #include "InvokeHTTP.h"
+#include "processors/ListenHTTP.h"
 #include "core/FlowFile.h"
-#include "../../unit/ProvenanceTestHelper.h"
+#include "unit/ProvenanceTestHelper.h"
 #include "core/Processor.h"
 #include "core/ProcessContext.h"
 #include "core/ProcessSession.h"
