+++ libminifi/test/unit/ProcessorTests.cpp
@@ -24,7 +24,6 @@
 #include <fstream>
 
 #include "../TestBase.h"
-#include "processors/ListenHTTP.h"
 #include "processors/LogAttribute.h"
 #include "processors/GetFile.h"
 #include "../unit/ProvenanceTestHelper.h"
