+++ libminifi/test/unit/PutFileTests.cpp
@@ -27,7 +27,6 @@
 
 #include "utils/file/FileUtils.h"
 #include "../TestBase.h"
-#include "processors/ListenHTTP.h"
 #include "processors/LogAttribute.h"
 #include "processors/GetFile.h"
 #include "processors/PutFile.h"
